c18n: Use a lock-free stack to hold a dynamic library's stacks
Previously, a lock is created for each dynamic library to guard access
to the stack that holds its stacks. But the number of locks that can be
dynamically created is in fact bounded, so a more scalable solution is
needed.

Fortunately, the stack is insert-only, allowing a simple lock-free
implementation.