Fixed the corprocess switcher on purecap kernels
The least significant bit of the capability that points to the switcher needs
to be set. Otherwise, we leave C64 mode after jumping into the switcher with
BLRS. The switcher code of purecap kernels uses capability instructions.
Attempting to execute these outside C64 results in an tag-violation error.

This commit also fixes two uses of non-capability instructions where the
switcher must use capabilties.